user_cleanup_script updated for mysql backends

+++ script/openchange_user_cleanup.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+# encoding: utf8
 import imaplib
 import optparse
 import MySQLdb
@@ -13,21 +14,34 @@ from samba.samdb import SamDB
 from samba.auth import system_session
 
 
-class SambaHelper(object):
+class SambaOCHelper(object):
     def __init__(self):
         self.samba_lp = LoadParm()
         self.samba_lp.set('debug level', '0')
-        self.samdb = SamDB(url=self.samdb_path(), lp=self.samba_lp,
+        self.samba_lp.load_default()
+        self.samdb = SamDB(url=self.samba_lp.private_path("sam.ldb"),
+                           lp=self.samba_lp,
                            session_info=system_session())
+        self.conn = self._open_mysql_connection()
 
-    def mapistore_folder(self):
-        return self.samba_lp.private_path("mapistore")
+    def _open_mysql_connection(self):
+        connection_string = self.samba_lp.get('mapiproxy:openchangedb')
+        if not connection_string:
+            raise Exception("Not found mapiproxy:openchangedb on samba configuration")
+        # mysql://openchange:password@localhost/openchange
+        m = re.search("(.+)://(.+):(.+)@(.+)/(.+)", connection_string)
+        if not m:
+            raise Exception("Unable to parse mapiproxy:openchangedb: %s" %
+                            connection_string)
+        if m.group(1) != 'mysql':
+            raise Exception("mapiproxy:openchangedb should start with mysql:// (we got %s)")
 
-    def samdb_path(self):
-        return self.samba_lp.private_path("sam.ldb")
+        dbuser = m.group(2)
+        dbpass = m.group(3)
+        dbhost = m.group(4)
+        dbname = m.group(5)
 
-    def openchange_ldb_path(self):
-        return self.samba_lp.private_path("openchange.ldb")
+        return MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)
 
     def invalid_user(self, username):
         ret = self.samdb.search(base=self.samdb.domain_dn(),
@@ -42,14 +56,13 @@ class SambaHelper(object):
         return ret[0]["mail"][0]
 
     def active_openchange_users(self):
-        conn = ldb.Ldb(self.openchange_ldb_path())
-        entries = conn.search(None, expression="(MailboxGUID=*)", attrs=["cn"],
-                              scope=ldb.SCOPE_SUBTREE)
-        return [entry["cn"][0] for entry in entries]
+        c = self.conn.cursor()
+        c.execute("SELECT name FROM mailboxes")
+        return sorted([row[0] for row in c.fetchall()])
 
 
 class ImapCleaner(object):
-    def __init__(self, imap_host="127.0.0.1", imap_port=143, dry_run=False, samba_helper=SambaHelper()):
+    def __init__(self, imap_host="127.0.0.1", imap_port=143, dry_run=False, samba_helper=SambaOCHelper()):
         self.imap_host = imap_host
         self.imap_port = imap_port
         self.dry_run = dry_run
@@ -58,6 +71,10 @@ class ImapCleaner(object):
     def cleanup(self, username, password=""):
         print "===== IMAP cleanup ====="
         client = imaplib.IMAP4(self.imap_host, self.imap_port)
+        if not password:
+            master_file = self.samba_helper.samba_lp.private_path('mapistore/master.password')
+            with open(master_file) as f:
+                password = f.read()
         try:
             email = self.samba_helper.find_email_of(username)
             code, data = client.login(email, password)
@@ -66,7 +83,8 @@ class ImapCleaner(object):
 
             print " [IMAP] Logged in as %s" % email
 
-            patterns = ["*(1)*", "Sync Issues*", "Spam*", "Problemas de sincroni*"]
+            patterns = ["*(1)*", "Spam", "Sync Issues*", "Problemas de sincroni*",
+                        "Problèmes de synchronisation*"]
             for pattern in patterns:
                 code, data = client.list("", pattern)
                 if code != "OK":
@@ -75,7 +93,16 @@ class ImapCleaner(object):
                     print " [IMAP] No folders with %s pattern" % pattern
                 else:
                     for d in data:
-                        folder_name = d.split("\"")[-2]
+                        folder_name = None
+                        m = re.match(".*\".\" \"(.*)\"", d)
+                        if m:
+                            folder_name = m.group(1)
+                        if not folder_name:
+                            m = re.match(".*\".\" (.*)", d)
+                            if m:
+                                folder_name = m.group(1)
+                        if not folder_name:
+                            raise Exception("Couldn't parse folder name on %r" % d)
                         if not self.dry_run:
                             code, data = client.delete(folder_name)
                             if code != "OK":
@@ -86,34 +113,34 @@ class ImapCleaner(object):
 
 
 class OpenchangeCleaner(object):
-    def __init__(self, samba_helper=SambaHelper(), dry_run=False):
+    def __init__(self, samba_helper=SambaOCHelper(), dry_run=False):
         self.samba_helper = samba_helper
         self.dry_run = dry_run
 
     def mapistore_indexing_cleanup(self, username):
         print "===== Mapistore indexing cleanup ====="
 
-        mapistore_user_dir = os.path.join(self.samba_helper.mapistore_folder(), username)
-        print " [Indexing] Deleting %s" % mapistore_user_dir
         if not self.dry_run:
-            shutil.rmtree(mapistore_user_dir, ignore_errors=True)
+            c = self.samba_helper.conn.cursor()
+            sql = ("DELETE FROM mapistore_indexes WHERE username = '%s'" %
+                   MySQLdb.escape_string(username))
+            c.execute(sql)
+            sql = ("DELETE FROM mapistore_indexing WHERE username = '%s'" %
+                   MySQLdb.escape_string(username))
+            c.execute(sql)
+            c.close()
+        print " [OC-Indexing] Deleted indexing database for %s" % username
 
     def openchangedb_cleanup(self, username):
         print "===== OpenchangeDB cleanup ====="
 
-        conn = ldb.Ldb(self.samba_helper.openchange_ldb_path())
-        entries = conn.search(None, expression="(cn=%s)" % (username),
-                              scope=ldb.SCOPE_SUBTREE)
-        if len(entries) != 1:
-            raise Exception(" [OpenchangeDB LDB] cn = %s must return 1 entry "
-                            "instead of %d" % (username, len(entries)))
-        subentries = conn.search(entries[0].dn.extended_str(),
-                                 expression="(distinguishedName=*)",
-                                 scope=ldb.SCOPE_SUBTREE)
-        for subentry in subentries:
-            if not self.dry_run:
-                conn.delete(subentry.dn)
-            print " [OpenchangeDB LDB] Deleted %s" % (subentry.dn)
+        if not self.dry_run:
+            c = self.samba_helper.conn.cursor()
+            sql = ("DELETE FROM mailboxes WHERE name = '%s'" %
+                   MySQLdb.escape_string(username))
+            c.execute(sql)
+            c.close()
+        print " [OC-OpenchangeDB] Deleted openchangedb for %s" % username
 
     def cleanup(self, username):
         self.openchangedb_cleanup(username)
@@ -121,54 +148,41 @@ class OpenchangeCleaner(object):
 
 
 class SOGoCleaner(object):
-    def __init__(self, dry_run=False, samba_helper=SambaHelper()):
+    def __init__(self, dry_run=False, samba_helper=SambaOCHelper()):
         self.dry_run = dry_run
         self.samba_helper = samba_helper
         self.system_defaults_file = "/etc/sogo/sogo.conf"
         self.user_defaults_file = os.path.expanduser("~sogo/GNUstep/Defaults/.GNUstepDefaults")
 
-    def asCSSIdentifier(self, inputString):
-        cssEscapingCharMap = {"_" : "_U_",
-                              "." : "_D_",
-                              "#" : "_H_",
-                              "@" : "_A_",
-                              "*" : "_S_",
-                              ":" : "_C_",
-                              "," : "_CO_",
-                              " " : "_SP_",
-                              "'" : "_SQ_",
-                              "&" : "_AM_",
-                              "+" : "_P_"}
-        newChars = []
-        for c in inputString:
-            if c in cssEscapingCharMap:
-                newChars.append(cssEscapingCharMap[c])
+    def _as_css_id(self, input_string):
+        css_char_map = {"_" : "_U_",
+                        "." : "_D_",
+                        "#" : "_H_",
+                        "@" : "_A_",
+                        "*" : "_S_",
+                        ":" : "_C_",
+                        "," : "_CO_",
+                        " " : "_SP_",
+                        "'" : "_SQ_",
+                        "&" : "_AM_",
+                        "+" : "_P_"}
+        new_chars = []
+        for c in input_string:
+            if c in css_char_map:
+                new_chars.append(css_char_map[c])
             else:
-                newChars.append(c)
-        return "".join(newChars)
+                new_chars.append(c)
+        return "".join(new_chars)
 
     def sogo_mysql_cleanup(self, dbhost, dbport, dbuser, dbpass, dbname, username):
         conn = MySQLdb.connect(host=dbhost, port=int(dbport), user=dbuser, passwd=dbpass, db=dbname)
         c = conn.cursor()
-        tablename = "socfs_%s" % self.asCSSIdentifier(username)
+        tablename = "sogo_cache_folder_%s" % self._as_css_id(username)
         if not self.dry_run:
             c.execute("TRUNCATE TABLE %s" % tablename)
         print " [SOGo MySQL] Table %s deleted" % tablename
 
-        """
-        c.execute("SELECT c_location, c_quick_location, c_acl_location "
-                  "FROM sogo_folder_info "
-                  "WHERE c_path2 = '%s'" % username.replace("'", "\\'"))
-        results = c.fetchall()
-        for tables in results:
-            for table in tables:
-                t = table.split('/')[-1]
-                if not self.dry_run:
-                    c.execute("TRUNCATE TABLE %s" % t)
-                print " [SOGo MySQL] Table %s deleted" % t
-        """
-
-    def _getOCSFolderInfoURL(self):
+    def _get_OCSFolderInfoURL(self):
         OCSFolderInfoURL = ""
         # read defaults from defaults files
         # order is important, user defaults must have precedence
@@ -184,9 +198,9 @@ class SOGoCleaner(object):
 
     def cleanup(self, username):
         print "===== SOGo cleanup ====="
-        OCSFolderInfoURL = self._getOCSFolderInfoURL()
+        OCSFolderInfoURL = self._get_OCSFolderInfoURL()
         if OCSFolderInfoURL is None:
-            raise Exception("ERROR Couldn't fetch OCSFolderInfoURL")
+            raise Exception("Couldn't fetch OCSFolderInfoURL")
 
         # mysql://sogo:sogo@127.0.0.1:5432/sogo/sogo_folder_info
         m = re.search("(.+)://(.+):(.+)@(.+):(\d+)/(.+)/(.+)", OCSFolderInfoURL)
@@ -223,7 +237,7 @@ if __name__ == "__main__":
     parser.add_option("--dry-run", action="store_true", help="Do not perform any action")
     opts, args = parser.parse_args()
 
-    samba_helper = SambaHelper()
+    samba_helper = SambaOCHelper()
 
     if opts.users:
         list_users(samba_helper)
